+++ tests/scripts/test_sync_library.py
@@ -1,9 +1,12 @@
 import json
 import re
 from pathlib import Path
+from tempfile import mkdtemp
+from unittest.mock import patch
 
 import pytest
 import requests_mock
+import typer
 
 from sekoia_automation.scripts.sync_library import SyncLibrary
 
@@ -250,47 +253,84 @@ manifests/{module['version']}"
 
 
 @requests_mock.Mocker(kw="m")
-def test_registry_check_custom_fail(module, **kwargs):
+def test_registry_check_auth_failure(module, **kwargs):
+    custom_path = "foo.bar"
+    custom_pathinfo = "v2"
+    image_name = module["docker"]
+    module["docker"] = f"{custom_path}/{custom_pathinfo}/{image_name}"
+
+    kwargs["m"].register_uri("GET", f"https://{custom_path}/token", status_code=403)
+    with pytest.raises(typer.Exit):
+        SyncLibrary(
+            SYMPOHNY_URL, API_KEY, Path("tests/data"), PAT, USER
+        ).check_image_on_registry(module["docker"], module["version"])
+
+    history = kwargs["m"].request_history
+    assert len(history) == 1
+    assert history[0].method == "GET"
+    assert history[0].url.startswith(f"https://{custom_path}/token")
+
+
+@requests_mock.Mocker(kw="m")
+def test_registry_check_not_found(module, **kwargs):
     custom_path = "foo.bar"
     custom_pathinfo = "v2"
     image_name = module["docker"]
     module["docker"] = f"{custom_path}/{custom_pathinfo}/{image_name}"
 
     kwargs["m"].register_uri(
-        "GET",
-        re.compile(f"https://{custom_path}.*"),
-        status_code=404,
-        json={"token": API_KEY},
+        "GET", f"https://{custom_path}/token", status_code=200, json={"token": API_KEY}
     )
-    assert not SyncLibrary(
-        SYMPOHNY_URL, API_KEY, Path("tests/data"), PAT, USER
-    ).check_image_on_registry(module["docker"], module["version"])
+    kwargs["m"].register_uri(
+        "GET",
+        f"https://{custom_path}/v2/sekoia-automation-module-sample/manifests/0.1",
+        status_code=404,
+    )
+    assert (
+        SyncLibrary(
+            SYMPOHNY_URL, API_KEY, Path("tests/data"), PAT, USER
+        ).check_image_on_registry(module["docker"], module["version"])
+        is False
+    )
 
     history = kwargs["m"].request_history
     assert len(history) == 2
     assert history[0].method == "GET"
     assert history[0].url.startswith(f"https://{custom_path}/token")
-    assert history[1].method == "GET"
-    assert (
-        history[1].url
-        == f"https://{custom_path}/{custom_pathinfo}/{image_name}/\
-manifests/{module['version']}"
+
+
+def test_check_image_no_token():
+    with pytest.raises(typer.Exit):
+        SyncLibrary(
+            SYMPOHNY_URL, API_KEY, Path("tests/data"), None, None, registry_check=True
+        ).execute()
+
+
+def test_get_module_logo():
+    lib = SyncLibrary(
+        SYMPOHNY_URL, API_KEY, Path("tests/data"), None, None, registry_check=True
     )
+    path = Path(mkdtemp())
+    assert lib.get_module_logo(path) is None
+
+    svg = path.joinpath("logo.svg")
+    svg.touch()
+    assert "svg+xml;base64" in lib.get_module_logo(path)
+    svg.unlink()
+
+    png = path.joinpath("logo.png")
+    png.touch()
+    assert "png;base64" in lib.get_module_logo(path)
+    png.unlink()
 
 
-@requests_mock.Mocker(kw="m")
-def test_registry_check_no_user(module, **kwargs):
-    assert SyncLibrary(
-        SYMPOHNY_URL, API_KEY, Path("tests/data"), registry_pat=PAT
-    ).check_image_on_registry(module["docker"], module["version"])
-    history = kwargs["m"].request_history
-    assert len(history) == 0
-
-
-@requests_mock.Mocker(kw="m")
-def test_registry_check_no_pwd(module, **kwargs):
-    assert SyncLibrary(
-        SYMPOHNY_URL, API_KEY, Path("tests/data"), registry_user=USER
-    ).check_image_on_registry(module["docker"], module["version"])
-    history = kwargs["m"].request_history
-    assert len(history) == 0
+def test_load_module_docker_image_not_found():
+    lib = SyncLibrary(
+        SYMPOHNY_URL, API_KEY, Path("tests/data"), None, None, registry_check=True
+    )
+    with patch(
+        "sekoia_automation.scripts.sync_library.SyncLibrary.check_image_on_registry"
+    ) as mock:
+        mock.return_value = False
+        with pytest.raises(typer.Exit):
+            lib.load_module(Path("tests/data/sample_module"))