+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add support for TLS client authentication in S3 storage
+- Launch the metrics exporter when the trigger starts
 
 ## [1.2.0] - 2023-05-10