+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add support for TLS client authentication in S3 storage
 - Launch the metrics exporter when the trigger starts
+- Wait for pod to be stopped by the API on critical error
 
 ## [1.2.0] - 2023-05-10