+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- In `synchronize-lib` script:
+    - Specify the registry in the image name
+    - Improve checking the existence of the image in the registry 
+
 ## [1.3.1] - 2023-06-06
 
 ### Changed