+++ CHANGELOG.md
@@ -7,13 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.2] - 2023-06-09
+
 ### Changed
 
 - In `synchronize-lib` script:
     - Specify the registry in the image name
     - Improve checking the existence of the image in the registry 
 
-## [1.3.1] - 2023-06-06
+## [1.3.1] - 2023-06-08
 
 ### Changed